GEODE-7150: Use Long in OQL COUNT and AVG (#3994)
- Use 'long' instead of 'int' to compute OQL COUNT result.
- Use 'long' instead of 'int' to compute OQL AVG intermediate results.

+++ geode-core/src/test/java/org/apache/geode/cache/query/internal/aggregate/AvgBucketNodeTest.java
@@ -68,7 +68,7 @@ public class AvgBucketNodeTest {
 
     Object result = avgBucketNode.terminate();
     assertThat(result).isInstanceOf(Object[].class);
-    assertThat(((Integer) ((Object[]) result)[0]).intValue()).isEqualTo(7);
+    assertThat(((Long) ((Object[]) result)[0]).intValue()).isEqualTo(7);
     assertThat(((Number) ((Object[]) result)[1]).intValue()).isEqualTo(28);
   }
 }
